Update webview functions
